extract colrev-internal-package discovery
